Implemented game simulation

+++ src/ColleyMatrix/Model/Team.cs
@@ -5,6 +5,6 @@
         public int TeamId { get; set; }
         public int Wins { get; set; }
         public int Losses { get; set; }
-        public float ColleyRating { get; set; }
+        public decimal ColleyRating { get; set; }
     }
 }
